+++ back-end/src/main/java/edu/fpt/ofbs/models/IDishDTO.java
@@ -1,11 +1,11 @@
 package edu.fpt.ofbs.models;
 
 public interface IDishDTO {
-	int getId();
-	
-	 String getDish_name();
-	
-	 String getImage_dish_id();
-	
-	 float getPrice();
+	long getId();
+
+	String getDish_name();
+
+	String getImage_dish_id();
+
+	float getPrice();
 }
