+++ back-end/src/main/java/edu/fpt/ofbs/repositories/PromotionRepository.java
@@ -10,7 +10,7 @@ import edu.fpt.ofbs.entities.Promotion;
 import edu.fpt.ofbs.models.IPromotionDTO;
 
 @Repository
-public interface PromotionRepository extends JpaRepository<Promotion, Integer>{
+public interface PromotionRepository extends JpaRepository<Promotion, Long>{
 	@Query(value = "exec sp_searchPromotionByProvince @province = ?1", nativeQuery = true)
 	List<IPromotionDTO> searchPromotionByProvince(String province);
 }
