+++ back-end/src/main/java/edu/fpt/ofbs/service/ComboService.java
@@ -14,7 +14,7 @@ public class ComboService {
 	@Autowired
 	private ComboRepository comboRepository;
 	
-	public List<IComboDTO> getCombosByRestaurantId(int restaurantId){
+	public List<IComboDTO> getCombosByRestaurantId(long restaurantId){
 		return comboRepository.getCombosByRestaurantId(restaurantId);
 	}
 }
